#286: Refactor for future expandability
