--- src/frontend/src/components/gameMap/effects/index.ts
@@ -1,3 +1,0 @@
-export { EffectContainer, type EffectContainerHandle } from './EffectContainer';
-export { EffectSprite } from './EffectSprite';
-export { EffectBatch } from './EffectBatch';
