+++ src/frontend/src/components/gameMap/particleEffects/index.ts
@@ -1,0 +1,3 @@
+export { EffectContainer, type EffectContainerHandle } from "./EffectContainer";
+export { EffectSprite } from "./support/EffectSprite";
+export { EffectBatch } from "./support/EffectBatch";
